z_projectmanagement/gdocs/pyproject.toml
[project]
name = "gdocs"
version = "0.1.0"
description = "Google Docs downloader utility"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.176.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
]